Select piece length when none is provided
When no piece length is provided to `imdl torrent create`, a piece
length is selected based on the size of the input. The hueristic is
lifted directly from libtorrent.

Also adds a `imdl torrent piece-length` command, which prints a table
of the piece lengths chosen at different content sizes, which is useful
for understanding and debugging the piece length selection algorithm.