Fix publish recipe
Previously, the release branch would be merged after pushing a tag to
github. This would cause the merge to fail, because github would kick
off a CI run after pushing the tag, and block the merge because it
wasn't finished.

This changes the publish recipe to merge before pushing the tag, which
aboves the above problem.

type: development